+++ templates/mobile/src/index.tsx
@@ -98,9 +98,7 @@ export function App() {
 export function mount({ container }) {
   ReactDOM.render(
     <HashRouter>
-      <StyleProvider hashPriority="high">
-        <App />
-      </StyleProvider>
+      <App />
     </HashRouter>,
     container
   );