+++ packages/vitdoc-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vitdoc/ui
 
+## 1.1.19
+
+### Patch Changes
+
+- fix: property style
+
 ## 1.1.18
 
 ### Patch Changes